front end routes set up and login form linked to auto scroll for home page
