--- .vscode/settings.json
@@ -1,19 +1,0 @@
-{
-    "workbench.colorCustomizations": {
-        "activityBar.activeBackground": "#b46a53",
-        "activityBar.activeBorder": "#275933",
-        "activityBar.background": "#b46a53",
-        "activityBar.foreground": "#e7e7e7",
-        "activityBar.inactiveForeground": "#e7e7e799",
-        "activityBarBadge.background": "#275933",
-        "activityBarBadge.foreground": "#e7e7e7",
-        "statusBar.background": "#945440",
-        "statusBar.foreground": "#e7e7e7",
-        "statusBarItem.hoverBackground": "#b46a53",
-        "titleBar.activeBackground": "#945440",
-        "titleBar.activeForeground": "#e7e7e7",
-        "titleBar.inactiveBackground": "#94544099",
-        "titleBar.inactiveForeground": "#e7e7e799"
-    },
-    "peacock.color": "#945440"
-}
