+++ package.json
@@ -14,6 +14,7 @@
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^3.4.3",
+    "react-scroll": "^1.8.1",
     "react-spring": "^8.0.27",
     "socket.io-client": "^2.3.0",
     "styled-components": "^5.1.1"
