+++ src/pages/MapDashboard/MapDashboard.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import './mapdashboard.css'
+
+export default function MapDashboard() {
+    return (
+        <div>
+            
+        </div>
+    )
+}
